+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implements the Bernoulli distribution in the PRG class
 - Implements `AttributeRange::next()` and `AttributeRange::prev()`
 - Plugins can now be reloaded
+- Main GUI: Adds a status bar and a console to show the error messages
 - Projects page: Adds button to maximize graph widgets
 - Projects page: Adds button to pause all experiments
 - Experiments table: Allows hiding columns